requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
databases
python-dotenv